Defaulted to Tranquility
- moved builder into crest
- moved baseuris to constants
