+++ eveapi/src/main/java/com/ccpgames/eveapi/Client/Crest.java
@@ -6,6 +6,7 @@ import com.google.gson.Gson;
 import java.util.HashMap;
 
 import retrofit.Callback;
+import retrofit.RestAdapter;
 import retrofit.RetrofitError;
 import retrofit.client.Response;
 import retrofit.http.GET;
@@ -17,13 +18,15 @@ import retrofit.http.QueryMap;
  */
 public class Crest {
 
-    Endpoint crestEndpoint;
+    public static final String SINGULARITY = "http://public-crest-sisi.testeveonline.com/";
+    public static final String TRANQUILITY = "https://crest-tq.eveonline.com/";
 
-    public Crest() {
-        this.crestEndpoint = new CrestBuilder()
-                .setSingularityEndpoint()
-                .build()
-                .crestEndpoint;
+    Endpoint crestEndpoint = new CrestBuilder()
+            .setTranquilityEndpoint()
+            .buildEndpoint();
+
+    public void setCrestEndpoint(Endpoint crestEndpoint) {
+        this.crestEndpoint = crestEndpoint;
     }
 
     interface Endpoint {
@@ -47,5 +50,35 @@ public class Crest {
                 callback.failure(error);
             }
         });
+    }
+
+    public static class CrestBuilder {
+
+        private String basePath;
+
+        public CrestBuilder setSingularityEndpoint() {
+            this.basePath = SINGULARITY;
+            return this;
+        }
+
+        public CrestBuilder setTranquilityEndpoint() {
+            this.basePath = TRANQUILITY;
+            return this;
+        }
+
+        public Crest build() {
+            Crest crest = new Crest();
+            crest.setCrestEndpoint(buildEndpoint());
+            return crest;
+        }
+
+        public Endpoint buildEndpoint() {
+            RestAdapter restAdapter = new RestAdapter.Builder()
+                    .setConverter(new StringConverter())
+                    .setEndpoint(basePath)
+                    .build();
+
+            return restAdapter.create(Endpoint.class);
+        }
     }
 }
