--- eveapi/src/main/java/com/ccpgames/eveapi/Client/CrestBuilder.java
@@ -1,33 +1,0 @@
-package com.ccpgames.eveapi.Client;
-
-import retrofit.RestAdapter;
-
-/**
- * Created by Jeremy Shore on 4/9/2015.
- */
-public class CrestBuilder {
-
-    private String basePath;
-
-    public CrestBuilder setSingularityEndpoint() {
-        this.basePath = "http://public-crest-sisi.testeveonline.com/";
-        return this;
-    }
-
-    public CrestBuilder setTranquilityEndpoint() {
-        this.basePath = "https://crest-tq.eveonline.com/";
-        return this;
-    }
-
-    public Crest build() {
-        Crest crest = new Crest();
-
-        RestAdapter restAdapter = new RestAdapter.Builder()
-                .setConverter(new StringConverter())
-                .setEndpoint(basePath)
-                .build();
-
-        crest.crestEndpoint = restAdapter.create(Crest.Endpoint.class);
-        return crest;
-    }
-}
